# scripts/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "Houses of the Oireachtas Open Data APIs",
...
    "license": {
      "name": "Oireachtas (Open Data) PSI Licence ",
      "url": "https://www.oireachtas.ie/en/open-data/license/"
    },
    "contact": {
      "name": "Oireachtas Open Data",
      "email": "open.data@oireachtas.ie",
      "url": "https://www.oireachtas.ie/en/open-data"
    }
  },
  "host": "api.oireachtas.ie",
  "basePath": "/v1",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/members": {
      "get": {
        "summary": "Members Endpoint",
        "description": "Returns a list of members.\nThe list supports paging.\n\n#### Mapping\n* date_start (Greater Than or Equal To) - member.memberships.membership.dateRange.start\n* date_end (Less Than or Equal To) - member.memberships.membership.dateRange.start\n* chamber_id - member.memberships.membership.house.uri\n* chamber - member.memberships.membership.house.houseCode\n* house_no - member.memberships.membership.house.houseNo\n* member_id - member.uri\n* skip - this will ignore the first x number of records set in the parameter\n* limit - this will only return a specific amount of records\n* party_code - member.memberships.membership.parties.party.partyCode\n* party_id - member.memberships.membership.parties.party.uri\n* const_code - member.memberships.membership.represents.represent.representCode\n* const_id - member.memberships.membership.represents.represent.uri\n* fuzzy_name_search - member.fullName",
        "tags": [
          "members"
        ],
        "operationId": "members",
        "parameters": [
          {
            "$ref": "#/parameters/dateStart"
          },
          {
            "$ref": "#/parameters/dateEnd"
          },
          {
            "$ref": "#/parameters/chamberID"
          },
          {
            "$ref": "#/parameters/house"
          },
          {
            "$ref": "#/parameters/houseNo"
          },
          {
            "$ref": "#/parameters/memberID"
          },
          {
            "$ref": "#/parameters/skip"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/partyCode"
          },
          {
            "$ref": "#/parameters/partyID"
          },
          {
            "$ref": "#/parameters/constituencyCode"
          },
          {
            "$ref": "#/parameters/constituencyID"
          },
          {
            "$ref": "#/parameters/fuzzyNameSearch"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "head": {
                  "type": "object",
                  "properties": {
                    "counts": {
                      "type": "object",
                      "properties": {
                        "memberCount": {
                          "type": "integer",
                          "example": 1927
                        },
                        "resultCount": {
                          "type": "integer",
                          "example": 1927
                        }
                      }
                    }
                  }
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "member": {
                        "type": "object",
                        "properties": {
                          "gender": { "type": "string" },
                          "uri": { "type": "string", "format": "uri" },
                          "pId": { "type": "string" },
                          "firstName": { "type": "string" },
                          "lastName": { "type": "string" },
                          "image": { "type": "boolean" },
                          "wikiTitle": { "type": "string" },
                          "memberships": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "membership": {
                                  "type": "object",
                                  "properties": {
                                    "house": {
                                      "type": "object",
                                      "properties": {
                                        "uri": { "type": "string", "format": "uri" },
                                        "houseCode": { "type": "string" },
                                        "houseNo": { "type": "string" },
                                        "chamberType": { "type": "string" },
                                        "showAs": { "type": "string" }
                                      }
                                    },
                                    "represents": {
                                      "type": "array",
                                      "items": {
                                        "type": "object",
                                        "properties": {
                                          "represent": {
                                            "type": "object",
                                            "properties": {
                                              "representCode": { "type": "string" },
                                              "uri": { "type": "string", "format": "uri" },
                                              "representType": { "type": "string" },
                                              "showAs": { "type": "string" }
                                            }
                                          }
                                        }
                                      }
                                    },
                                    "uri": { "type": "string", "format": "uri" },
                                    "dateRange": {
                                      "type": "object",
                                      "properties": {
                                        "start": { "type": "string", "format": "date" },
                                        "end": { "type": "string", "format": "date" }
                                      }
                                    },
                                    "committees": {
                                      "type": "array",
                                      "items": {}
                                    },
                                    "parties": {
                                      "type": "array",
                                      "items": {
                                        "type": "object",
                                        "properties": {
                                          "party": {
                                            "type": "object",
                                            "properties": {
                                              "partyCode": { "type": "string" },
                                              "uri": { "type": "string", "format": "uri" },
                                              "dateRange": {
                                                "type": "object",
                                                "properties": {
                                                  "start": { "type": "string", "format": "date" },
                                                  "end": { "type": "string", "format": "date" }
                                                }
                                              },
                                              "showAs": { "type": "string" }
                                            }
                                          }
                                        }
                                      }
                                    },
                                    "offices": {
                                      "type": "array",
                                      "items": {}
                                    }
                                  }
                                }
                              }
                            }
                          },
                          "dateOfDeath": { "type": "string", "format": "date" },
                          "fullName": { "type": "string" },
                          "memberCode": { "type": "string" },
                          "showAs": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
  "parameters": {
...
    "memberID": {
      "name": "member_id",
      "in": "query",
      "description": "Filter by Member uri, e.g. /ie/oireachtas/member/id/Brendan-Howlin.S.1983-02-23",
      "required": false,
      "type": "string"
    },
...
    "house": {
      "name": "chamber",
      "in": "query",
      "description": "Filter by House name (dail or seanad).\nUsing an empty string retrieves results for both Houses.",
      "required": false,
      "type": "string",
      "default": "",
      "enum": [
        "",
        "dail",
        "seanad"
      ]
    },
    "chamberID": {
      "name": "chamber_id",
      "in": "query",
      "description": "Filter by house or committee uri, e.g. /ie/oireachtas/house/dail/32",
      "required": false,
      "type": "array",
      "collectionFormat": "csv",
      "allowEmptyValue": true,
      "default": []
    },
    "dateStart": {
      "name": "date_start",
      "in": "query",
      "description": "In the format YYYY-MM-DD (e.g. 1900-12-31)",
      "required": false,
      "type": "string",
      "format": "date"
    },
    "dateEnd": {
      "name": "date_end",
      "in": "query",
      "description": "In the format YYYY-MM-DD (e.g. 2025-12-31)",
      "required": false,
      "type": "string",
      "format": "date"
    },
...
    "skip": {
      "name": "skip",
      "in": "query",
      "description": "This allows skipping of records by a specific integer.",
      "required": false,
      "type": "integer",
      "default": 0
    },
    "limit": {
      "name": "limit",
      "in": "query",
      "description": "This allows the limiting of records to a specific integer.",
      "required": false,
      "type": "integer",
      "default": 50
    },
    "houseNo": {
      "name": "house_no",
      "in": "query",
      "description": "Filter by house number (integer only)",
      "required": false,
      "type": "integer"
    },
...
    "partyCode": {
      "name": "party_code",
      "in": "query",
      "description": "Filter by party code, e.g. Labour_Party",
      "required": false,
      "type": "string"
    },
    "partyID": {
      "name": "party_id",
      "in": "query",
      "description": "Filter by unique party identifier (uri), e.g. /ie/oireachtas/party/dail/32/Labour_Party",
      "required": false,
      "type": "string"
    },
    "constituencyCode": {
      "name": "const_code",
      "in": "query",
      "description": "Filter by constituency code, e.g. Administrative-Panel",
      "required": false,
      "type": "string"
    },
    "constituencyID": {
      "name": "const_id",
      "in": "query",
      "description": "Filter by unique constituency identifier (uri), e.g. /ie/oireachtas/house/seanad/26/panel/Administrative-Panel",
      "required": false,
      "type": "string"
    },
    "fuzzyNameSearch": {
      "name": "fuzzy_name_search",
      "in": "query",
      "description": "Filter by name with fuzzy search (typo safe)",
      "required": false,
      "type": "string"
    },
...
  }
}